Adjustment aws s3 service and add s3 faker server
