+++ faker-s3/.gitignore
@@ -1,0 +1,20 @@
+# compiled output
+/dist
+/node_modules
+
+yarn.lock
+
+# Logs
+logs
+*.log
+npm-debug.log*
+pnpm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+lerna-debug.log*
+
+# OS
+.DS_Store
+
+# Tests
+/tmp
